typing: make the project pyright-clean (standard) and enforce it (#1784)
Bring the whole repo (pymobiledevice3/ + tests/) to zero pyright errors
in standard mode and wire it into pre-commit + CI so it stays there.

Approach: real fixes over suppressions. Honest Optional annotations,
explicit narrowing that raises proper exceptions where the old code
would have crashed, raising accessors for set-once-on-connect attributes,
@overloads where optionality depends on an argument, and required
parameters where a value is never actually None. Only rule-specific
`# pyright: ignore[...]` remain, for inherently dynamic APIs (pyusb,
pykdebugparser, construct, platform-gated constants).

Enforcement:
- [tool.pyright] in pyproject: typeCheckingMode=standard, pythonVersion
  3.9, include pymobiledevice3/tests, venvPath/venv pinned to .venv.
- local pyright pre-commit hook pinned to 1.1.411; a dedicated macos/
  py3.14 CI job runs the same hook after installing the project (the
  test matrix skips it via SKIP=pyright).
- ruff UP007/UP045 ignored so Union/Optional aren't rewritten to X | Y /
  X | None — the 3.9 floor needs annotations evaluable under
  typing.get_type_hints(), not merely string-safe.
- typing_extensions becomes a direct dependency (dataclass_transform,
  Self); CLI choices use native typer Enums (dropping a click_type shim).

Real runtime bugs surfaced and fixed along the way:
- mobile_activation called the nonexistent _request_async()
- restore: missing await on get_device_generated_firmware_data() (SE and
  Rose iOS 18+ firmware paths)
- tss add_timer_tags passed the parameters dict as its own lookup key
- web_protocol: Page.stopScreencast called a nonexistent ScreenCast
  method; Image.ANTIALIAS was removed in Pillow 10
- get_remoted_devices used a sync `with` on an async-only context manager
- installation_proxy.upgrade() leaked *args into a bool parameter
- webinspector built a message payload as a dead tuple and never sent it

Contract cleanups: raising accessors (RemoteService.service, Tap.channel,
ASRClient.service, RemoteXPCConnection/RemotePairingTunnelService reader/
writer, IRecv device/mode) replace repeated `assert x is not None`;
specific exceptions (NotConnectedError/NotPairedError/AppInstallError/
MissingValueError/DvtException) replace the generic base; try_decode
gains a str-returning overload; get_personalized_data/stitch_component
take a required TSS; bonjour enforces the SRV port at construction; the
duplicated _current_task_name helper is consolidated into utils.

+++ pymobiledevice3/services/bt_packet_logger.py
@@ -84,7 +84,7 @@ async def write_packetlogger_stream(out: BinaryIO, packet_generator: AsyncIterab
 
 
 async def write_pcapng_stream(
-    out: BinaryIO, packet_generator: AsyncIterable[bytes], product_version: str = "", tz_offset_seconds: int = 0
+    out: BinaryIO, packet_generator: AsyncIterable[bytes], product_version: str = "", tz_offset_seconds: float = 0
 ) -> None:
     # The BTPacketLogger service reports record timestamps as device-local wall-clock time, whereas the pcapng
     # EnhancedPacket timestamp (with the default if_tsresol of 1e-6) is interpreted as UTC. Subtract the device's
